Shrink DaemonClient toward deletion (Phase 3)
Extract non-class code from DaemonClient.swift into dedicated files
(VellumPaths, UsageModels, DaemonNotifications), remove three broken
send() callers that silently fell through the dispatcher chain, delete
all six no-op dispatcher files, and replace the generic dispatch loop
with a direct UserMessageMessage type check.
