+++ clients/macos/vellum-assistant/App/AppDelegate+Notifications.swift
@@ -309,18 +309,10 @@ extension AppDelegate {
         errorCode: String?
     ) {
         guard let deliveryId else { return }
-        let result = NotificationIntentResult(
-            type: "notification_intent_result",
-            deliveryId: deliveryId,
-            success: success,
-            errorMessage: errorMessage,
-            errorCode: errorCode
-        )
-        do {
-            try daemonClient.send(result)
-        } catch {
-            log.warning("Failed to send notification_intent_result for deliveryId \(deliveryId): \(error.localizedDescription)")
-        }
+        // No-op: notification intent result acking only worked with the legacy
+        // WebSocket transport. The NotificationIntentResult message is no longer
+        // handled by any dispatcher or REST endpoint.
+        log.debug("sendNotificationIntentResult: skipped for deliveryId \(deliveryId) — no REST endpoint available")
     }
 
     func deliverNotificationIntent(_ msg: NotificationIntentMessage) {
