+++ clients/shared/Network/DaemonClient.swift
@@ -9,47 +9,6 @@ private let networkLog = OSLog(
     category: .pointsOfInterest
 )
 
-/// Resolve the `.vellum` data directory, honoring `BASE_DATA_DIR` when set.
-public func resolveVellumDir(environment: [String: String]? = nil) -> String {
-    let env = environment ?? ProcessInfo.processInfo.environment
-    if let baseDir = env["BASE_DATA_DIR"]?.trimmingCharacters(in: .whitespacesAndNewlines), !baseDir.isEmpty {
-        let resolved = baseDir == "~" ? NSHomeDirectory() : (baseDir.hasPrefix("~/") ? NSHomeDirectory() + "/" + String(baseDir.dropFirst(2)) : baseDir)
-        return resolved + "/.vellum"
-    }
-    // Check the lockfile for instance-specific directory (multi-instance support)
-    if let instanceDir = resolveInstanceDirFromLockfile() {
-        return instanceDir + "/.vellum"
-    }
-    return NSHomeDirectory() + "/.vellum"
-}
-
-/// Read the instanceDir from the latest lockfile entry's resources.
-private func resolveInstanceDirFromLockfile() -> String? {
-    guard let json = LockfilePaths.read(),
-          let assistants = json["assistants"] as? [[String: Any]],
-          !assistants.isEmpty else {
-        return nil
-    }
-    // Find the most recently hatched entry
-    let sorted = assistants.sorted { a, b in
-        let dateA = a["hatchedAt"] as? String ?? ""
-        let dateB = b["hatchedAt"] as? String ?? ""
-        return dateA > dateB
-    }
-    guard let latest = sorted.first,
-          let resources = latest["resources"] as? [String: Any],
-          let instanceDir = resources["instanceDir"] as? String,
-          !instanceDir.isEmpty else {
-        return nil
-    }
-    return instanceDir
-}
-
-/// Resolve the daemon PID file path, honoring `BASE_DATA_DIR`.
-public func resolvePidPath(environment: [String: String]? = nil) -> String {
-    return resolveVellumDir(environment: environment) + "/vellum.pid"
-}
-
 /// Protocol for daemon client communication, enabling dependency injection and testing.
 @MainActor
 public protocol DaemonClientProtocol {
@@ -60,134 +19,6 @@ public protocol DaemonClientProtocol {
     func disconnect()
     func startSSE()
     func stopSSE()
-}
-
-// MARK: - Usage Response Models
-
-/// Aggregate totals for a time range from `GET /v1/usage/totals`.
-public struct UsageTotalsResponse: Decodable, Equatable, Sendable {
-    public let totalInputTokens: Int
-    public let totalOutputTokens: Int
-    public let totalCacheCreationTokens: Int
-    public let totalCacheReadTokens: Int
-    public let totalEstimatedCostUsd: Double
-    public let eventCount: Int
-    public let pricedEventCount: Int
-    public let unpricedEventCount: Int
-
-    public init(
-        totalInputTokens: Int,
-        totalOutputTokens: Int,
-        totalCacheCreationTokens: Int,
-        totalCacheReadTokens: Int,
-        totalEstimatedCostUsd: Double,
-        eventCount: Int,
-        pricedEventCount: Int,
-        unpricedEventCount: Int
-    ) {
-        self.totalInputTokens = totalInputTokens
-        self.totalOutputTokens = totalOutputTokens
-        self.totalCacheCreationTokens = totalCacheCreationTokens
-        self.totalCacheReadTokens = totalCacheReadTokens
-        self.totalEstimatedCostUsd = totalEstimatedCostUsd
-        self.eventCount = eventCount
-        self.pricedEventCount = pricedEventCount
-        self.unpricedEventCount = unpricedEventCount
-    }
-}
-
-/// A single day bucket from `GET /v1/usage/daily`.
-public struct UsageDayBucket: Decodable, Equatable, Sendable {
-    public let date: String
-    public let totalInputTokens: Int
-    public let totalOutputTokens: Int
-    public let totalEstimatedCostUsd: Double
-    public let eventCount: Int
-
-    public init(date: String, totalInputTokens: Int, totalOutputTokens: Int, totalEstimatedCostUsd: Double, eventCount: Int) {
-        self.date = date
-        self.totalInputTokens = totalInputTokens
-        self.totalOutputTokens = totalOutputTokens
-        self.totalEstimatedCostUsd = totalEstimatedCostUsd
-        self.eventCount = eventCount
-    }
-}
-
-/// Response wrapper for `GET /v1/usage/daily`.
-public struct UsageDailyResponse: Decodable, Equatable, Sendable {
-    public let buckets: [UsageDayBucket]
-
-    public init(buckets: [UsageDayBucket]) {
-        self.buckets = buckets
-    }
-}
-
-/// A single grouped breakdown row from `GET /v1/usage/breakdown`.
-public struct UsageGroupBreakdownEntry: Decodable, Equatable, Sendable {
-    public let group: String
-    public let totalInputTokens: Int
-    public let totalOutputTokens: Int
-    public let totalCacheCreationTokens: Int
-    public let totalCacheReadTokens: Int
-    public let totalEstimatedCostUsd: Double
-    public let eventCount: Int
-
-    public init(
-        group: String,
-        totalInputTokens: Int,
-        totalOutputTokens: Int,
-        totalCacheCreationTokens: Int = 0,
-        totalCacheReadTokens: Int = 0,
-        totalEstimatedCostUsd: Double,
-        eventCount: Int
-    ) {
-        self.group = group
-        self.totalInputTokens = totalInputTokens
-        self.totalOutputTokens = totalOutputTokens
-        self.totalCacheCreationTokens = totalCacheCreationTokens
-        self.totalCacheReadTokens = totalCacheReadTokens
-        self.totalEstimatedCostUsd = totalEstimatedCostUsd
-        self.eventCount = eventCount
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case group
-        case totalInputTokens
-        case totalOutputTokens
-        case totalCacheCreationTokens
-        case totalCacheReadTokens
-        case totalEstimatedCostUsd
-        case eventCount
-    }
-
-    public init(from decoder: any Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        group = try container.decode(String.self, forKey: .group)
-        totalInputTokens = try container.decode(Int.self, forKey: .totalInputTokens)
-        totalOutputTokens = try container.decode(Int.self, forKey: .totalOutputTokens)
-        totalCacheCreationTokens = try container.decodeIfPresent(Int.self, forKey: .totalCacheCreationTokens) ?? 0
-        totalCacheReadTokens = try container.decodeIfPresent(Int.self, forKey: .totalCacheReadTokens) ?? 0
-        totalEstimatedCostUsd = try container.decode(Double.self, forKey: .totalEstimatedCostUsd)
-        eventCount = try container.decode(Int.self, forKey: .eventCount)
-    }
-}
-
-/// Response wrapper for `GET /v1/usage/breakdown`.
-public struct UsageBreakdownResponse: Decodable, Equatable, Sendable {
-    public let breakdown: [UsageGroupBreakdownEntry]
-
-    public init(breakdown: [UsageGroupBreakdownEntry]) {
-        self.breakdown = breakdown
-    }
-}
-
-extension Notification.Name {
-    /// Posted by `DaemonClient` on the main actor immediately after `isConnected` transitions to `true`.
-    public static let daemonDidReconnect = Notification.Name("daemonDidReconnect")
-
-    /// Posted when the daemon's signing key fingerprint changes, indicating an instance switch.
-    /// Observers should trigger credential re-bootstrap.
-    public static let daemonInstanceChanged = Notification.Name("daemonInstanceChanged")
 }
 
 /// Platform-agnostic client for communicating with the Vellum daemon via HTTP + SSE.
