--- clients/shared/Network/HTTPDaemonClient+Apps.swift
@@ -1,20 +1,0 @@
-import Foundation
-import os
-
-private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant", category: "HTTPTransport")
-
-// MARK: - Apps Domain Dispatcher
-
-extension HTTPTransport {
-
-    func registerAppsRoutes() {
-        registerDomainDispatcher { message in
-            if message is AppDataRequestMessage {
-                // Handled by AppsClient via GatewayHTTPClient.
-                return true
-            }
-
-            return false
-        }
-    }
-}
